clean(frontend): rename some classes
